refactor: 게임을 진행하는 main 함수는 별도로 분리한다
