+++ README.md
@@ -156,6 +156,7 @@
 #### [Todo List]
 
 - [ ] 피드백 반영
+  - [x] 게임을 진행하는 main 함수는 별도로 분리한다.
 - [ ] domain 패키지와 view 패키지로 각각 분리하고, 의존성을 제한한다.
 - [ ] 테스트 불가능한 영역을 분리한다.
 - [ ] 테스트 가능한 모든 영역에 단위 테스트를 진행한다.
